requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
PyPDF2==3.0.1
Werkzeug==3.0.0
Jinja2==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6

